[android][core] Extract shared callback context from JavaCallback
