+++ packages/expo-modules-core/android/src/main/cpp/JavaCallback.cpp
@@ -8,23 +8,6 @@
 
 namespace expo {
 
-JavaCallback::CallbackContext::CallbackContext(
-  jsi::Runtime &rt,
-  std::weak_ptr<react::CallInvoker> jsCallInvokerHolder,
-  std::optional<jsi::Function> resolveHolder,
-  std::optional<jsi::Function> rejectHolder
-) : react::LongLivedObject(rt),
-    rt(rt),
-    jsCallInvokerHolder(std::move(jsCallInvokerHolder)),
-    resolveHolder(std::move(resolveHolder)),
-    rejectHolder(std::move(rejectHolder)) {}
-
-void JavaCallback::CallbackContext::invalidate() {
-  resolveHolder.reset();
-  rejectHolder.reset();
-  allowRelease();
-}
-
 JavaCallback::JavaCallback(std::shared_ptr<CallbackContext> callbackContext)
   : callbackContext(std::move(callbackContext)) {}
 
@@ -64,40 +47,17 @@ jni::local_ref<JavaCallback::javaobject> JavaCallback::newInstance(
 void JavaCallback::invokeWithResolver(
   std::function<void(jsi::Runtime &rt, jsi::Function &jsFunction)> resolver
 ) {
-  const auto strongCallbackContext = this->callbackContext.lock();
-  // The context were deallocated before the callback was invoked.
-  if (strongCallbackContext == nullptr) {
-    return;
-  }
-
-  const auto jsInvoker = strongCallbackContext->jsCallInvokerHolder.lock();
-  // Call invoker is already released, so we cannot invoke the callback.
-  if (jsInvoker == nullptr) {
-    return;
-  }
-
-  jsInvoker->invokeAsync(
-    [
-      context = callbackContext,
-      resolver = std::move(resolver)
-    ]() -> void {
-      auto strongContext = context.lock();
-      // The context were deallocated before the callback was invoked.
-      if (strongContext == nullptr) {
-        return;
-      }
-
-      if (!strongContext->resolveHolder.has_value()) {
+  scheduleOnJSThread(
+    callbackContext,
+    [resolver = std::move(resolver)](jsi::Runtime &rt, CallbackContext &context) {
+      if (!context.resolveHolder.has_value()) {
         throw std::runtime_error(
           "JavaCallback was already settled. Cannot invoke it again"
         );
       }
 
-      jsi::Function &jsFunction = strongContext->resolveHolder.value();
-      jsi::Runtime &rt = strongContext->rt;
-
-      resolver(rt, jsFunction);
-      strongContext->invalidate();
+      resolver(rt, context.resolveHolder.value());
+      context.invalidate();
     });
 }
 
@@ -259,38 +219,17 @@ void JavaCallback::invokeFloatArray(jni::alias_ref<jni::JArrayFloat> result) {
 }
 
 void JavaCallback::invokeError(jni::alias_ref<jstring> code, jni::alias_ref<jstring> errorMessage) {
-  const auto strongCallbackContext = this->callbackContext.lock();
-  // The context were deallocated before the callback was invoked.
-  if (strongCallbackContext == nullptr) {
-    return;
-  }
-
-  const auto jsInvoker = strongCallbackContext->jsCallInvokerHolder.lock();
-  // Call invoker is already released, so we cannot invoke the callback.
-  if (jsInvoker == nullptr) {
-    return;
-  }
-
-  jsInvoker->invokeAsync(
+  scheduleOnJSThread(
+    callbackContext,
     [
-      context = callbackContext,
       code = code->toStdString(),
       errorMessage = errorMessage->toStdString()
-    ]() -> void {
-      auto strongContext = context.lock();
-      // The context were deallocated before the callback was invoked.
-      if (strongContext == nullptr) {
-        return;
-      }
-
-      if (!strongContext->rejectHolder.has_value()) {
+    ](jsi::Runtime &rt, CallbackContext &context) {
+      if (!context.rejectHolder.has_value()) {
         throw std::runtime_error(
           "JavaCallback was already settled. Cannot invoke it again"
         );
       }
-
-      jsi::Function &jsFunction = strongContext->rejectHolder.value();
-      jsi::Runtime &rt = strongContext->rt;
 
       auto codedError = makeCodedError(
         rt,
@@ -298,13 +237,13 @@ void JavaCallback::invokeError(jni::alias_ref<jstring> code, jni::alias_ref<jstr
         jsi::String::createFromUtf8(rt, errorMessage)
       );
 
-      jsFunction.call(
+      context.rejectHolder->call(
         rt,
         (const jsi::Value *) &codedError,
         (size_t) 1
       );
 
-      strongContext->invalidate();
+      context.invalidate();
     });
 }
 } // namespace expo
